UPDATE - Matrices iteration + drop down menu
Changed how the program runs throught the vector, it is now using a random value to go around, but its too noisy.
Also changed the name of a few variables to make it more comprehensible.
Added a step to show to the user each color used and a drop down menu by the side of it to select which zone is going to be used to each props.

The noise and the menus still need a lot of improvement. Need to find a way to run throught the matrices in a more natural way. like from inwards to outwards. Maybe using more iterations of the process picking just the initial pixel coordenate in a random way